Generate portal levels at the same time as their source
Whenever a regular level has a portal, this will generate the portal
levels immediately after the regular level's generation is triggered.
This imposes a deterministic generation order on portal levels for
incremental pregen, as well as a deterministic order between portals
when there are multiple ones from the same level (which is missing from
even full pregen before this commit).

This commit also puts ziggurats in the generation order, though they are
handled a bit differently; this is so everything else generates before
they do, to try to minimize the impacts of item placement in zigs.