Don't pregenerate bazaars (for now at least)
The problem is that multiple bazaars can generate, which I hadn't
realized. It would be possible to pregenerate the first one without too
much trouble, but this leads to some oddness with the current
incremental pregen setup. Ideally, we would pregenerate all bazaars, but
this would require some tricky changes to how portal vaults are
currently handled.